requirements.txt
numpy==1.23.5
pandas==1.4.4
streamlit
pickle
matplotlib==3.6.2
seaborn==0.12.1
nltk==3.7
scikit-learn
pip==23.0.1
joblib==1.2.0
Pillow==9.3.0